Move previous files to genius package
