+++ genius/search.go
@@ -1,4 +1,4 @@
-package main
+package genius
 
 import (
 	"bufio"
